add: refactor ws
