+++ src/main.rs
@@ -309,7 +309,7 @@ async fn main() -> Result<(), anyhow::Error> {
             // init websocket gc
             if cfg.websocket.enabled {
                 log::info!("Initializing WebSocket session garbage collector");
-                if let Err(e) = handler::http::request::ws_v2::init().await {
+                if let Err(e) = handler::http::request::ws::init().await {
                     job_init_tx.send(false).ok();
                     panic!("websocket gc init failed: {}", e);
                 }
